This pull request represents some work I have been doing to support decoding in nnet3 in a way that naturally allows re-use of previously computed things.  This is done right inside the nnet3 compilation process, so we don't have to mess around externally with additional outputs and inputs.

It's applicable to decoding with TDNNs and LSTMs (and should be particularly helpful for LSTMs, because it avoids the need to extend the chunk with left context).
I haven't yet implemented the "online" decoding with this.. that can be done later.
The point this is at, is that all the C++ code is implemented, and tested based on automated tests, but I haven't modified any scripts or run any of the examples.
This is not just about testing new code paths, but existing code could in theory crash, because I made a lot of changes to core nnet3 stuff.
Because I have made the 'looped' computation available by adding a new binary (nnet3-latgen-faster-looped), nnet3 decoding-script changes would be required to make use of this: probably an option --looped-computation [true|false] that would trigger a branch in the script.  Note: the options --extra-left-context and --extra-right-context are not present in nnet3-latgen-faster-looped.  There is a config --extra-left-context-initial (extra left context for the first chunk), but I hope it won't be necessary to use or won't make much difference.

I am hoping someone can help me test this and do the script-level changes.  @tomkocse?  @vijayaditya, do you have any suggestion about this?