@vijayaditya 2016-11-14
I would recommend testing the following:
1. When the model is trained with a fixed left-context does decoding with an infinite left-context help ?
     (I think it might. During my aspire experiments I once trained a BLSTM with 0 left/right context, due to a script error. When decoding using an extra-{left,right}-context of 40 was important even if there was a mismatch with the training. Using 0 extra-{left,right}-context led to an increase WER by 10% rel. IIRC)

2. This is slightly related to this PR so I am mentioning it here:
Do we need to change the training strategy by using variable chunk-left-context during training ? @GaofengCheng 's dropout in recurrent connections would be useful here. But he would have to update the dropout component to use the same dropout mask value for the entire history vector rather than a separate mask value for each element of the history vector. Also as we want to ensure that we simulate infinite left context during training we might want to check if larger chunk-left-context values would help (they might not; previous experiments did not show any change in results when using contexts of 40,60 or 80).

Finally as @freewym implemented the context preserving LSTM, I think he might benefit by following this PR closely, as he can check what changed in this implementation and how is better/worse compared to his implementation. @freewym Also, let us know if you can guide @tomkocse in designing the tests.

I will add more comments when I go through the PR.