@tomkocse 2016-11-15
I could help if a slight delay is tolerable .