@danpovey 2016-11-15
OK.

On Mon, Nov 14, 2016 at 10:03 PM, Tom Ko <notifications@github.com> wrote:

> I could help if a slight delay is tolerable .
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/kaldi-asr/kaldi/pull/1186#issuecomment-260534186>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/ADJVu5ZHFLg8KLmE6LQyeWixwZltC3Hbks5q-SDxgaJpZM4Kw5uH>
> .
>