@GaofengCheng 2016-11-23
@vijayaditya do you mean I should update a dropout component (maybe we can name it ScaleDropoutComponent) like this:
    if scaledropout-rate=0.2 then a matrix like this is added: ` [[0.8,0.8...]...[0.8,0.8,...]]`
My experiments using the DropoutComponent accomplished by @pegahgh improves results on Chain BLSTM AMI SDM.
If the ScaleDropoutComponent is what you mean.
I think we can do experiments between the ScaleDropoutComponent and DropoutComponent