@danpovey 2016-11-23
No, he means a form of dropout where instead of the dropout-mask matrix having random zeros or ones, it has rows that are randomly either zero or one [but the same for each row.]