@tomkocse 2016-11-23
@Dan, I start to work on this PR. To make sure i am working in a right way, here are my procedures:
1. check out the nnet3 binaries in branch danpovey:looped_computation
2. use the TDNN+LSTM scripts to train the nnet model
3. test this nnet model with either [true|false] in the --looped-computation option
Is this fine ?