@danpovey 2016-11-23
Yes, the model is unchanged.
It's about using the new decoding program with -looped in the name.
If you have any C++-level crashes, no need to debug it yourself, just ask
me.



On Wed, Nov 23, 2016 at 2:58 AM, Vijayaditya Peddinti <
notifications@github.com> wrote:

> If I understand correctly you have to just test a change during decoding.
> You need not train a new model. You can use the models available
> at /export/b01/vpeddinti/xconfig/egs/swbd/s5c/exp/chain/tdnn_
> lstm_1a_ld5_sp.
>
> --Vijay
>
> On Wed, Nov 23, 2016 at 2:47 AM, Tom Ko <notifications@github.com> wrote:
>
> > @Dan <https://github.com/Dan>, I start to work on this PR. To make sure
> i
> > am working in a right way, here are my procedures:
> >
> > 1. check out the nnet3 binaries in branch danpovey:looped_computation
> > 2. use the TDNN+LSTM scripts to train the nnet model
> > 3. test this nnet model with either [true|false] in the
> > --looped-computation option
> > Is this fine ?
> >
> > —
> > You are receiving this because you were mentioned.
> > Reply to this email directly, view it on GitHub
> > <https://github.com/kaldi-asr/kaldi/pull/1186#issuecomment-262450710>,
> or mute
> > the thread
> > <https://github.com/notifications/unsubscribe-auth/
> ADtwoKcaba6VHUzD1ALV4nucuGiDL6R7ks5rA-9_gaJpZM4Kw5uH>
> > .
>
> >
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/kaldi-asr/kaldi/pull/1186#issuecomment-262452101>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/ADJVu-dbHUo8BR_hmLUr6GVDgZTvKCO9ks5rA_IbgaJpZM4Kw5uH>
> .
>