@GaofengCheng 2016-11-23
@danpovey @vijayaditya I plan to compare the two dropout methods, my experiments show dropout by Pegah have already helped, but we should try more methods. I will update a new dropout in this week, @danpovey  renaming the dropout by Pegah DotDropout and naming the idea by Vijay RowDropout, do you think it is OK?