@danpovey 2016-11-23
Conventional dropout is what we have implemented, so the name Dropout is
probably suitable.  Anyway, it would probably be easiest to implement this
as an option to the DropoutComponent, as it only really affects one line of
the code, where the mask is created.  Maybe an option 'dropout_per_frame'
(default = false).

Dan




On Wed, Nov 23, 2016 at 6:46 PM, Gaofeng Cheng <notifications@github.com>
wrote:

> @danpovey <https://github.com/danpovey> @vijayaditya
> <https://github.com/vijayaditya> I plan to compare the two dropout
> methods, my experiments show dropout by Pegah have already helped, but we
> should try more methods. I will update a new dropout in this week,
> @danpovey <https://github.com/danpovey> renaming the dropout by Pegah
> DotDropout and naming the idea by Vijay RowDropout, do you think it is OK?
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/kaldi-asr/kaldi/pull/1186#issuecomment-262657166>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/ADJVu_vrytgpukhQSET9GLOfic_ledvpks5rBNBBgaJpZM4Kw5uH>
> .
>