@tomkocse 2016-11-26
@danpovey i remove the use of option --extra-left-context-initial in he decoding and i get an error 
ERROR (nnet3-latgen-faster-looped:CreateComputation():nnet-compile.cc:58) Not all outputs were computable, cannot create computation.