@tomkocse 2016-11-26
the original decoding script was using  --extra-left-context-initial=-1 but nnet3-latgen-faster-looped:Check() not allow to use this option with negative value