@tomkocse 2016-11-29
Here are the real time factors for 5 out of 50 decoding processes:
without looped:
LOG (nnet3-latgen-faster:main():nnet3-latgen-faster.cc:246) Time taken 150.483s: real-time factor assuming 100 frames/sec is 0.448454
LOG (nnet3-latgen-faster:main():nnet3-latgen-faster.cc:246) Time taken 164.609s: real-time factor assuming 100 frames/sec is 0.899796
LOG (nnet3-latgen-faster:main():nnet3-latgen-faster.cc:246) Time taken 97.1057s: real-time factor assuming 100 frames/sec is 0.974077
LOG (nnet3-latgen-faster:main():nnet3-latgen-faster.cc:246) Time taken 283.33s: real-time factor assuming 100 frames/sec is 0.820485
LOG (nnet3-latgen-faster:main():nnet3-latgen-faster.cc:246) Time taken 239.113s: real-time factor assuming 100 frames/sec is 0.897536

with looped:
LOG (nnet3-latgen-faster-looped:main():nnet3-latgen-faster-looped.cc:251) Time taken 312.587s: real-time factor assuming 100 frames/sec is 0.931537
LOG (nnet3-latgen-faster-looped:main():nnet3-latgen-faster-looped.cc:251) Time taken 226.34s: real-time factor assuming 100 frames/sec is 1.23723
LOG (nnet3-latgen-faster-looped:main():nnet3-latgen-faster-looped.cc:251) Time taken 66.7531s: real-time factor assuming 100 frames/sec is 0.669607
LOG (nnet3-latgen-faster-looped:main():nnet3-latgen-faster-looped.cc:251) Time taken 405.598s: real-time factor assuming 100 frames/sec is 1.17456
LOG (nnet3-latgen-faster-looped:main():nnet3-latgen-faster-looped.cc:251) Time taken 162s: real-time factor assuming 100 frames/sec is 0.608084

@vijayaditya I will use your TDNN+LSTM model to decode